refactor: move ScreenReaderOnly into SharedComponents
