+++ src/components/NewButtons/NewButtonLabelWithScreenReaderTitle.tsx
@@ -3,8 +3,9 @@ import styled from "styled-components";
 
 import typography from "../../styles/utils/typography";
 import { Span } from "../Typography";
-import ScreenReaderOnly from "../ScreenReaderOnly/ScreenReaderOnly";
 import { SpanProps } from "../Typography/Span";
+
+import ScreenReaderOnly from "@/components/SharedComponents/ScreenReaderOnly";
 
 export type ButtonLabelProps = {
   children: React.ReactNode;
